Basic extraction into factory functions
