+++ packages/use-intl/src/core/index.tsx
@@ -1,0 +1,9 @@
+export {default as AbstractIntlMessages} from './AbstractIntlMessages';
+export {
+  default as TranslationValues,
+  RichTranslationValues
+} from './TranslationValues';
+export {default as Formats} from './Formats';
+export {default as DateTimeFormatOptions} from './DateTimeFormatOptions';
+export {default as NumberFormatOptions} from './NumberFormatOptions';
+export {default as IntlError, IntlErrorCode} from './IntlError';
