+++ packages/use-intl/src/index.tsx
@@ -1,15 +1,2 @@
-export {default as IntlProvider} from './IntlProvider';
-export {default as AbstractIntlMessages} from './AbstractIntlMessages';
-export {default as useTranslations} from './useTranslations';
-export {
-  default as TranslationValues,
-  RichTranslationValues
-} from './TranslationValues';
-export {default as useIntl} from './useIntl';
-export {default as useLocale} from './useLocale';
-export {default as useNow} from './useNow';
-export {default as useTimeZone} from './useTimeZone';
-export {default as Formats} from './Formats';
-export {default as DateTimeFormatOptions} from './DateTimeFormatOptions';
-export {default as NumberFormatOptions} from './NumberFormatOptions';
-export {default as IntlError, IntlErrorCode} from './IntlError';
+export * from './core';
+export * from './react';
