+++ packages/use-intl/src/react/index.tsx
@@ -1,0 +1,6 @@
+export {default as IntlProvider} from './IntlProvider';
+export {default as useTranslations} from './useTranslations';
+export {default as useIntl} from './useIntl';
+export {default as useLocale} from './useLocale';
+export {default as useNow} from './useNow';
+export {default as useTimeZone} from './useTimeZone';
