+++ packages/use-intl/src/react/useTranslations.tsx
@@ -1,12 +1,14 @@
 import {ReactElement, ReactNodeArray} from 'react';
-import Formats from './Formats';
-import TranslationValues, {RichTranslationValues} from './TranslationValues';
+import Formats from '../core/Formats';
+import TranslationValues, {
+  RichTranslationValues
+} from '../core/TranslationValues';
+import MessageKeys from '../core/utils/MessageKeys';
+import NamespaceKeys from '../core/utils/NamespaceKeys';
+import NestedKeyOf from '../core/utils/NestedKeyOf';
+import NestedValueOf from '../core/utils/NestedValueOf';
 import useIntlContext from './useIntlContext';
 import useTranslationsImpl from './useTranslationsImpl';
-import MessageKeys from './utils/MessageKeys';
-import NamespaceKeys from './utils/NamespaceKeys';
-import NestedKeyOf from './utils/NestedKeyOf';
-import NestedValueOf from './utils/NestedValueOf';
 
 /**
  * Translates messages from the given namespace by using the ICU syntax.
